add "declassification assignments" for the result of the election
Based on the trick from the _hybrid approach_, assignments are added
to the code of the server, to make it explicit that the result of the
voting, although technically computed using the high variable,
actually depends only on low variables. Without it, Joana reports an
information flow. These assignments are justified using the KeY tool
(it is proved that they are redundant, i.e. do not change the state of
the system).