justification of "declassification assignments" for the result
Assignments added to the code verified by Joana need to be justified
by KeY. (Those assignments have to do with the result of the election
which, although from the point of view of Joana dependends on the high
variable, is actually only dependent on low variables).

Here we add the appropriate equations to be proven.