make the communication between the server and BB use AMT
The server used to send the result of the election to the bulletin
board over an SAMT channel (which provides secrecy and authentication).
Because secrecy is not required for this (and because Joana had some
problems with distinguishing two instances of SAMT), now we use the AMT
functionality (Authenticated Message Transmission) for this purpose.