Little change in Setup to please JOANA
With a construct like 'choice = secret ? choices0[i] : choices1[i]'
either the choices0 or the choices1 array are accessed, dependent on secret. As array accesses
may result in an exception, this exception dependends on the secret. With the proposed change,
we get both elements in advance, so that any exception happens independent of what the secret is.