Little change in Server to please JOANA
The int variable 'candidate' may depend on the secret choice. It is then used as an array index.
Array accesses may result in an index-out-of-exception. So, dependent on the secret choice, an
exception may occur and crash the whole program, so that certain public outputs do not happen.
With the proposed change, this is prevented.